Fixed some stuff, working on UI

+++ colors.h
@@ -5,5 +5,5 @@ struct Color{
 	uint8_t g;
 	uint8_t b;
 	Color();
-	void setHue(uint8_t hue);
+	void setHue(int hue);
 };
